Optimization update
1. resetting game function has been optimized. It still lags as the game
restarts so it's good to refresh the whole page once a while, but it
should be slightly better.

Shotgun buff:
since shotgun has random angle variance, it completely misses an obious
target. This is now changed so that at always least one bullet is shot
straight. It is more consistant even in long range.